test: introduce createWorkspaceContext
